Arrumando a estilização
