+++ modulo3/projeto-labeEdit/labe-edit/src/pages/ErrorPage.jsx
@@ -1,8 +1,13 @@
 import React from 'react'
-
+import Cat from '../assets/cat.gif'
  const ErrorPage = () => {
   return (
-    <div>ErrorPage</div>
+    <div>
+     <h1>
+     ErrorPage
+    </h1>
+    <img src={Cat} />
+    </div>
   )
 }
